adjust directory
