+++ paddle/fluid/pybind/pir.cc
@@ -1727,19 +1727,6 @@ void BindPassManager(pybind11::module *m) {
            [](PassManager &self) { self.EnableIRPrinting(); });
 }
 
-#ifdef PADDLE_WITH_DISTRIBUTE
-void BindDistOpsAPI(pybind11::module *module) {
-  {
-    if (PyModule_AddFunctions(module->ptr(), DistOpsAPI) < 0) {
-      {
-        PADDLE_THROW(
-            phi::errors::Fatal("Add C++ DistOpsAPI to core.ops failed!"));
-      }
-    }
-  }
-}
-#endif
-
 void BindPir(pybind11::module *module) {
   auto ir_module = module->def_submodule("pir");
   BindProgram(&ir_module);
@@ -1756,9 +1743,6 @@ void BindPir(pybind11::module *module) {
   BindControlFlowApi(&ir_module);
   auto ops_modules = ir_module.def_submodule("ops");
   BindOpsAPI(&ops_modules);
-#ifdef PADDLE_WITH_DISTRIBUTE
-  BindDistOpsAPI(&ops_modules);
-#endif
   BindIrParser(&ir_module);
 }
